move `get_solution`
