+++ cobra/core/solution.py
@@ -8,7 +8,12 @@ import logging
 from builtins import object, super
 from warnings import warn
 
-__all__ = ("Solution",)
+from numpy import zeros
+from pandas import Series
+
+from cobra.exceptions import OptimizationError
+
+__all__ = ("Solution", "LegacySolution", "get_solution")
 
 LOGGER = logging.getLogger(__name__)
 
@@ -253,3 +258,59 @@ class LegacySolution(object):
         """
         warn("unnecessary to call this deprecated function",
              DeprecationWarning)
+
+
+def get_solution(model, reactions=None, metabolites=None):
+    """
+    Generate a solution representation of the current solver state.
+
+    Paramters
+    ---------
+    model : cobra.Model
+        The model whose reactions to retrieve values for.
+    reactions : list, optional
+        An iterable of `cobra.Reaction` objects. Uses `model.reactions` by
+        default.
+    metabolites : list, optional
+        An iterable of `cobra.Metabolite` objects. Uses `model.metabolites` by
+        default.
+
+    Returns
+    -------
+    cobra.Solution
+
+    Note
+    ----
+    This is only intended for the `optlang` solver interfaces and not the legacy
+    solvers.
+    """
+    if model.solver.status != "optimal":
+        raise OptimizationError(
+            "The solver status is '{0:s}'. Cannot reliably retrieve values."
+            .format(model.solver.status))
+    if reactions is None:
+        reactions = model.reactions
+    if metabolites is None:
+        metabolites = model.metabolites
+
+    rxn_index = [rxn.id for rxn in reactions]
+    fluxes = zeros(len(reactions))
+    var_primals = model.solver.primal_values
+    reduced = zeros(len(reactions))
+    var_duals = model.solver.reduced_costs
+    for (i, rxn) in enumerate(reactions):
+        forward = rxn.forward_variable.name
+        reverse = rxn.reverse_variable.name
+        fluxes[i] = var_primals[forward] - var_primals[reverse]
+        reduced[i] = var_duals[forward] - var_duals[reverse]
+    met_index = [met.id for met in metabolites]
+    shadow = zeros(len(metabolites))
+    constr_duals = model.solver.shadow_prices
+    for (i, met) in enumerate(metabolites):
+        shadow[i] = constr_duals[met.id]
+    return Solution(model.solver.objective.value, model.solver.status,
+                    reactions,
+                    Series(index=rxn_index, data=fluxes),
+                    Series(index=rxn_index, data=reduced),
+                    metabolites,
+                    Series(index=met_index, data=shadow))
