+++ cobra/flux_analysis/loopless.py
@@ -7,11 +7,9 @@ from __future__ import absolute_import
 import numpy
 from six import iteritems
 from sympy.core.singleton import S
-
-from cobra.core import Metabolite, Reaction
+from cobra.core import Metabolite, Reaction, get_solution
 from cobra.util import add_to_solver, linear_reaction_coefficients
 from cobra.manipulation.modify import convert_to_irreversible
-from cobra.util.solver import get_solution
 
 try:
     import numpy
